Make the "replace dev with prod" script less prone to nuking PROD if the right variables aren't set and the pg_dump fails
Oops

+++ db-pg/scripts/replace-dev-with-prod.sh
@@ -8,14 +8,23 @@ chmod 600 ~/.pgpass
 prod_db_dump_file="digitalocean_db-a_${DB_SCHEMA}_${now}.dump"
 pg_dump --host=$DB_HOST --username=$DB_USERNAME --format=custom --schema=$DB_SCHEMA $DB_NAME > $prod_db_dump_file
 
+DB_HOST=
+DB_PORT=
+DB_NAME=
+DB_USERNAME=
+DB_PASSWORD=
 
-echo "Backup development database"
-export $(xargs < secrets/sausage-dev-db.env)
-echo "${DB_HOST}:${DB_PORT}:${DB_NAME}:${DB_USERNAME}:${DB_PASSWORD}" > ~/.pgpass
+# If prod dump exists and is not empty
+if [[ -f $prod_db_dump_file && -s $prod_db_dump_file ]]; then
+  echo "Backup development database"
+  export $(xargs < secrets/sausage-dev-db.env)
+  echo "${DB_HOST}:${DB_PORT}:${DB_NAME}:${DB_USERNAME}:${DB_PASSWORD}" > ~/.pgpass
 
-psql --host=$DB_HOST --username=$DB_USERNAME --dbname=$DB_NAME --command='DROP SCHEMA IF EXISTS "'"$DB_SCHEMA"'_bak" CASCADE;'
-psql --host=$DB_HOST --username=$DB_USERNAME --dbname=$DB_NAME --command='ALTER SCHEMA "'"$DB_SCHEMA"'" RENAME TO "'"$DB_SCHEMA"'_bak";'
+  psql --host=$DB_HOST --username=$DB_USERNAME --dbname=$DB_NAME --command='DROP SCHEMA IF EXISTS "'"$DB_SCHEMA"'_bak" CASCADE;'
+  psql --host=$DB_HOST --username=$DB_USERNAME --dbname=$DB_NAME --command='ALTER SCHEMA "'"$DB_SCHEMA"'" RENAME TO "'"$DB_SCHEMA"'_bak";'
 
+  echo "Load production database dump to development (host=$DB_HOST)"
+  pg_restore --host=$DB_HOST --username=$DB_USERNAME --dbname=$DB_NAME --single-transaction $prod_db_dump_file
+fi
 
-echo "Load production database dump to development"
-pg_restore --host=$DB_HOST --username=$DB_USERNAME --dbname=$DB_NAME --single-transaction $prod_db_dump_file
+echo "Fin"
